refactor: modalStyleファイルをlibディレクトリに移動
